// authServiceC/src/app/types/authService.pb.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.2
//   protoc               v3.21.12
// source: authService.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "AuthService";

export interface AuthRequest {
  token: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  verifiedAt: string;
}

export interface AuthResponse {
  success: boolean;
  message: string;
  user: User | undefined;
}

export const AUTH_SERVICE_PACKAGE_NAME = "AuthService";

export interface AuthServiceClient {
  authCheck(request: AuthRequest): Observable<AuthResponse>;
}

export interface AuthServiceController {
  authCheck(request: AuthRequest): Promise<AuthResponse> | Observable<AuthResponse> | AuthResponse;
}

export function AuthServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["authCheck"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("AuthService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("AuthService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const AUTH_SERVICE_NAME = "AuthService";